feat(ui): split dashboard monolithic hook and extract polling configs
