+++ .github/workflows/cd.yml
@@ -136,6 +136,9 @@ jobs:
       - name: Build frontend
         env:
           VITE_API_BASE: ${{ vars.VITE_API_BASE || secrets.VITE_API_BASE }}
+          VITE_DASHBOARD_REALTIME_POLL_MS: ${{ vars.VITE_DASHBOARD_REALTIME_POLL_MS || '5000' }}
+          VITE_DASHBOARD_STATE_POLL_MS: ${{ vars.VITE_DASHBOARD_STATE_POLL_MS || '15000' }}
+          VITE_DASHBOARD_CONFIG_POLL_MS: ${{ vars.VITE_DASHBOARD_CONFIG_POLL_MS || '60000' }}
         run: npm run build
 
       - name: Validate remote frontend directory
