+++ configs/app.example.env
@@ -30,3 +30,8 @@ BINANCE_FUTURES_KLINE_BASE_URL=https://fapi.binance.com
 BINANCE_FUTURES_WS_URL=wss://fstream.binance.com/ws
 # Optional OKX public market data websocket override.
 OKX_PUBLIC_WS_URL=wss://ws.okx.com:8443/ws/v5/public
+
+# Frontend Dashboard Polling Intervals
+VITE_DASHBOARD_REALTIME_POLL_MS=5000
+VITE_DASHBOARD_STATE_POLL_MS=15000
+VITE_DASHBOARD_CONFIG_POLL_MS=60000
