+++ web/console/.env.example
@@ -1,0 +1,4 @@
+VITE_API_BASE=http://localhost:8081
+VITE_DASHBOARD_REALTIME_POLL_MS=5000
+VITE_DASHBOARD_STATE_POLL_MS=15000
+VITE_DASHBOARD_CONFIG_POLL_MS=60000
